fix: solve code review requests
