+++ src/pages/index.tsx
@@ -1,29 +1,12 @@
-import { gql } from '@apollo/client';
 import { PokeApIqueryQueryResult } from 'generated/graphql';
 import { initializeApollo } from 'graphql/apollo-client';
+import { GET_POKEMONS, PAGE_SIZE } from 'queries';
 
 import { HomeTemplate } from 'Templates/Home';
 
 type HomeProps = {
   pokemonsData: PokeApIqueryQueryResult;
 };
-
-export const GET_POKEMONS = gql`
-  query pokeAPIquery($limit: Int!, $offset: Int!) {
-    pokemon_v2_pokemon(limit: $limit, offset: $offset) {
-      name
-      id
-      weight
-      height
-      pokemon_v2_pokemontypes {
-        pokemon_v2_type {
-          name
-          id
-        }
-      }
-    }
-  }
-`;
 
 export default function Home({ pokemonsData }: HomeProps) {
   return <HomeTemplate pokemons={pokemonsData} />;
@@ -35,7 +18,7 @@ export async function getServerSideProps() {
   const pokemonsData = await apolloClient.query({
     query: GET_POKEMONS,
     variables: {
-      limit: 9,
+      limit: PAGE_SIZE,
       offset: 0,
     },
   });
